chore: separate abstract test classes and unify header usage
Extract abstract test base classes into their own files and
replace manual .header() calls with .headers(getHeaders())
for consistent header management across test and IT classes.
