+++ apps/wanaku-router-backend/src/test/java/ai/wanaku/backend/api/v1/namespaces/NamespacesResourceTest.java
@@ -1,197 +1,13 @@
 package ai.wanaku.backend.api.v1.namespaces;
 
-import jakarta.ws.rs.core.MediaType;
-import jakarta.ws.rs.core.Response;
-
-import java.time.Instant;
-import java.util.HashMap;
-import java.util.Map;
-import org.jboss.logging.Logger;
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.TestProfile;
 import ai.wanaku.backend.support.NoOidcTestProfile;
-import ai.wanaku.backend.support.TestIndexHelper;
-import ai.wanaku.backend.support.WanakuRouterTest;
-import ai.wanaku.capabilities.sdk.api.types.Namespace;
 
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.CoreMatchers.nullValue;
-
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.MethodOrderer;
-import org.junit.jupiter.api.Order;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 @QuarkusTest
 @TestProfile(NoOidcTestProfile.class)
 public class NamespacesResourceTest extends AbstractNamespacesResourceTest {}
-
-abstract class AbstractNamespacesResourceTest extends WanakuRouterTest {
-    private static final Logger LOG = Logger.getLogger(NamespacesResourceTest.class);
-    private static final String TEST_LABEL_KEY = "testsuite";
-    private static final String TEST_LABEL_VALUE = "namespaces";
-
-    private static String createdId;
-    private static String createdPath;
-    private static String staleId;
-    private static String stalePath;
-
-    @BeforeAll
-    static void setup() {
-        TestIndexHelper.clearAllCaches();
-    }
-
-    protected java.util.Map<String, String> getHeaders() {
-        return java.util.Map.of("Content-Type", MediaType.APPLICATION_JSON);
-    }
-
-    @Order(1)
-    @Test
-    public void testCreateNamespace() {
-        Namespace namespace = new Namespace();
-        createdPath = "test-ns-" + System.currentTimeMillis();
-        namespace.setPath(createdPath);
-        Map<String, String> labels = new HashMap<>();
-        labels.put(TEST_LABEL_KEY, TEST_LABEL_VALUE);
-        namespace.setLabels(labels);
-
-        final io.restassured.response.Response response =
-                given().headers(getHeaders()).body(namespace).when().post("/api/v1/namespaces");
-
-        LOG.infof("Response: %s", response.getBody().asString());
-
-        createdId = response.then()
-                .statusCode(Response.Status.OK.getStatusCode())
-                .body("data.path", is(createdPath))
-                .body("data.labels." + TEST_LABEL_KEY, is(TEST_LABEL_VALUE))
-                .body("data.labels.'wanaku.io/preallocated'", is("true"))
-                .extract()
-                .path("data.id");
-    }
-
-    @Order(2)
-    @Test
-    public void testListNamespacesByLabel() {
-        given().headers(getHeaders())
-                .queryParam("labelFilter", TEST_LABEL_KEY + "=" + TEST_LABEL_VALUE)
-                .when()
-                .get("/api/v1/namespaces")
-                .then()
-                .statusCode(Response.Status.OK.getStatusCode())
-                .body(String.format("data.find { it.id == '%s' }.path", createdId), is(createdPath));
-    }
-
-    @Order(3)
-    @Test
-    public void testGetNamespaceById() {
-        given().headers(getHeaders())
-                .when()
-                .get("/api/v1/namespaces/" + createdId)
-                .then()
-                .statusCode(Response.Status.OK.getStatusCode())
-                .body("data.id", is(createdId))
-                .body("data.path", is(createdPath));
-    }
-
-    @Order(4)
-    @Test
-    public void testUpdateNamespace() {
-        Namespace namespace = new Namespace();
-        namespace.setId(createdId);
-        namespace.setPath(createdPath);
-        namespace.setName("allocated");
-        Map<String, String> labels = new HashMap<>();
-        labels.put(TEST_LABEL_KEY, TEST_LABEL_VALUE);
-        labels.put("wanaku.io/preallocated", "true");
-        namespace.setLabels(labels);
-
-        given().headers(getHeaders())
-                .body(namespace)
-                .when()
-                .put("/api/v1/namespaces/" + createdId)
-                .then()
-                .statusCode(Response.Status.OK.getStatusCode());
-
-        given().headers(getHeaders())
-                .when()
-                .get("/api/v1/namespaces/" + createdId)
-                .then()
-                .statusCode(Response.Status.OK.getStatusCode())
-                .body("data.name", is("allocated"));
-    }
-
-    @Order(5)
-    @Test
-    public void testDeleteNamespace() {
-        given().headers(getHeaders())
-                .when()
-                .delete("/api/v1/namespaces/" + createdId)
-                .then()
-                .statusCode(Response.Status.OK.getStatusCode());
-
-        given().headers(getHeaders())
-                .when()
-                .get("/api/v1/namespaces/" + createdId)
-                .then()
-                .statusCode(Response.Status.OK.getStatusCode())
-                .body("data.id", is(createdId))
-                .body("data.path", is(createdPath))
-                .body("data.name", nullValue())
-                .body("data.labels.'wanaku.io/preallocated'", is("true"));
-
-        given().headers(getHeaders())
-                .queryParam("labelFilter", TEST_LABEL_KEY + "=" + TEST_LABEL_VALUE)
-                .when()
-                .get("/api/v1/namespaces")
-                .then()
-                .statusCode(Response.Status.OK.getStatusCode())
-                .body(String.format("data.find { it.id == '%s' }", createdId), nullValue());
-    }
-
-    @Order(6)
-    @Test
-    public void testStaleNamespaceCleanup() {
-        Namespace namespace = new Namespace();
-        stalePath = "test-stale-ns-" + System.currentTimeMillis();
-        namespace.setPath(stalePath);
-        Map<String, String> labels = new HashMap<>();
-        labels.put("testsuite", "stale");
-        labels.put("wanaku.io/expires-at", String.valueOf(Instant.now().getEpochSecond() - 10));
-        namespace.setLabels(labels);
-
-        final io.restassured.response.Response response =
-                given().headers(getHeaders()).body(namespace).when().post("/api/v1/namespaces");
-
-        staleId = response.then()
-                .statusCode(Response.Status.OK.getStatusCode())
-                .extract()
-                .path("data.id");
-
-        given().headers(getHeaders())
-                .queryParam("maxAgeSeconds", 604800)
-                .when()
-                .get("/api/v1/namespaces/stale")
-                .then()
-                .statusCode(Response.Status.OK.getStatusCode())
-                .body(String.format("data.find { it.id == '%s' }.path", staleId), is(stalePath));
-
-        given().headers(getHeaders())
-                .queryParam("maxAgeSeconds", 0)
-                .when()
-                .delete("/api/v1/namespaces/stale")
-                .then()
-                .statusCode(Response.Status.OK.getStatusCode())
-                .body("data", is(1));
-
-        given().headers(getHeaders())
-                .queryParam("maxAgeSeconds", 604800)
-                .when()
-                .get("/api/v1/namespaces/stale")
-                .then()
-                .statusCode(Response.Status.OK.getStatusCode())
-                .body(String.format("data.find { it.id == '%s' }", staleId), nullValue());
-    }
-}
